Making the Main Window a different class
